projects/zakaz_2/requirements.txt
fastapi
uvicorn
pika
psycopg2-binary
sqlalchemy
yt-dlp
python-dotenv
opencv-python-headless
